+++ services/users/src/main/scala/api/mappers/TokenMappers.scala
@@ -1,0 +1,14 @@
+package api.mappers
+
+import api.dto.TokensDTO
+import model.UserTokens
+
+object TokenMappers:
+  extension (tokens: UserTokens)
+    def toTokensDTO: TokensDTO =
+      TokensDTO(
+        accessToken = tokens.accessToken,
+        expiresIn = tokens.expiresIn,
+        refreshToken = tokens.refreshToken,
+        refreshExpiresIn = tokens.refreshExpiresIn
+      )