+++ services/users/src/main/scala/model/RefreshRequest.scala
@@ -1,0 +1,9 @@
+package model
+
+import upickle.default.Reader
+import upickle.default.macroR
+
+case class RefreshRequest(refreshToken: String)
+
+object RefreshRequest:
+  given Reader[RefreshRequest] = macroR